+++ containers/daap-python-base/tests/unit/conftest.py
@@ -1,0 +1,4 @@
+import sys
+from os.path import dirname, join
+
+sys.path.append(join(dirname(__file__), "../", "../", "src", "var", "task"))